Update postman collection

+++ App/BackEnd/App.Content.Domain/Operations/Commands/DeleteArticle/DeleteArticleCommandHandler.cs
@@ -27,9 +27,17 @@ namespace App.Content.Domain.Operations.Commands.DeleteArticle
             if ( !await _articleRepository.ExistsBySlug(request.Slug) )
                 return OperationResponseFactory.NotFound<DeleteArticleCommand, OperationResponse<DeleteArticleCommandResult>>(typeof(ArticleEntity), request.Slug);
 
+            var article = await _articleRepository.GetBySlug(request.Slug, _userContext.UserId);
+            if ( article.Author.UserId != _userContext.UserId )
+                return OperationResponseFactory.NotAuthorized<DeleteArticleCommand, OperationResponse<DeleteArticleCommandResult>>("article does not belong to the user");
+
             await _articleRepository.Delete(_userContext.UserId, request.Slug);
 
-            return OperationResponseFactory.Success(new DeleteArticleCommandResult());
+            return OperationResponseFactory.Success(new DeleteArticleCommandResult
+            {
+                ArticleId = article.Id,
+                UserId = _userContext.UserId
+            });
         }
     }
 }
